+++ pytsite/routing/__init__.py
@@ -1,11 +1,10 @@
 """PytSite Routing
 """
-from . import _error as error
-from ._rule import Rule
-from ._rules_map import RulesMap
-from ._controller import Controller, ControllerArgs
-
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
+from . import _error as error
+from ._rule import Rule
+from ._rules_map import RulesMap
+from ._controller import ControllerArgs, Controller, Filter